+++ demo/src/components/Renderer.tsx
@@ -1,9 +1,24 @@
+import { useState } from 'react'
 import Canvas, { CanvasProps } from 'react-canvas-animate'
 
 interface RendererProps extends CanvasProps {}
 
 function Renderer({ ...props }: RendererProps) {
-  return <Canvas {...props} />
+  const [scroll, setScroll] = useState({ x: 0, y: 0 })
+
+  return (
+    <Canvas
+      onWheel={(e) =>
+        setScroll({
+          x: scroll.x - e.deltaX / 2,
+          y: scroll.y - e.deltaY / 2,
+        })
+      }
+      gridOffset={scroll}
+      gridSize={40}
+      {...props}
+    />
+  )
 }
 
 export default Renderer